generating a random multiplication table
