+++ package.json
@@ -7,6 +7,8 @@
 		"@testing-library/react": "^9.3.2",
 		"@testing-library/user-event": "^7.1.2",
 		"@types/axios": "^0.14.0",
+		"@types/cors": "^2.8.7",
+		"@types/express": "^4.17.8",
 		"@types/jest": "^24.0.0",
 		"@types/node": "^12.0.0",
 		"@types/react": "^16.9.0",
@@ -29,11 +31,12 @@
 		"redux-logger": "^3.0.6",
 		"redux-thunk": "^2.3.0",
 		"styled-components": "^5.2.0",
+		"ts-node": "^9.0.0",
 		"typescript": "~3.7.2"
 	},
 	"scripts": {
 		"start": "react-scripts start",
-		"dev": "concurrently \"nodemon server/index.js\" \"react-scripts start\"",
+		"dev": "concurrently \"nodemon src/server/index.js\" \"react-scripts start\"",
 		"build": "react-scripts build",
 		"test": "react-scripts test",
 		"eject": "react-scripts eject",
