+++ src/styles/QuestionCard.styles.ts
@@ -2,6 +2,7 @@ import styled from 'styled-components';
 
 export const Wrapper = styled.div`
 	max-width: 1100px;
+	min-width: 320px;
 	background: #ebfeff;
 	border-radius: 10px;
 	border: 2px solid #0085a3;
